updated uml for TradeInParser
changed parameter type from Set to List
changed return type of startTrade() and calculateNumNewPieces() from boolean to int, so that it can be used to update the amount of armies the players have.

Concerns: relationship and interaction of this class with other classes is still a little unclear and not sure how the new armies will be updated in other classes. However, for now, will just leave it as returning an int value and passing into other classes. Could be modified later if find a better design.